requirements.txt
# PTB-XL Edge ECG Challenge — Requirements
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Data science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# ECG / Signal
wfdb>=4.1.0

# Image processing (for image_preprocessing module)
opencv-python>=4.8.0

# ONNX export & runtime
onnx>=1.15.0
onnxruntime>=1.16.0     # Use onnxruntime for standard; onnxruntime-arm for RPi builds

# Config
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional: for mixed-precision training on CUDA
# torch[cu118]  # or your CUDA version

# Raspberry Pi (install on RPi 5 only):
# onnxruntime==1.16.3   # ARM wheel